+++ app/overrides/projects/show.rb
@@ -1,0 +1,5 @@
+Deface::Override.new virtual_path: 'projects/show',
+                     name: 'view-projects-show-bottom-hook',
+                     insert_after: 'div.splitcontentright',
+                     original: '8cd4d1b38e8afcb4665dbfea661b7048fbd92cf7',
+                     partial: 'hooks/view_projects_show_bottom'