+++ assets/javascripts/noreferrer.js
@@ -1,6 +1,0 @@
-$(function() {
-  var handleNoReferrer = function() {
-    this.rel = 'noreferrer';
-  };
-  $('a.external').each(handleNoReferrer);
-});